api/mongo_crud/requirements.txt
flask-restplus==0.13.0
Flask-SQLAlchemy==2.1
numpy==1.16.4
pandas==0.25.3
joblib==0.14.1
pymongo==3.9.0
Werkzeug==0.16.0
